requirements.txt
ultralytics
opencv-python-headless
streamlit
streamlit-webrtc
av
pillow
numpy
torch
